requirements-osbs.txt
# See below for guidance and scripts to assist in generating the full dependencies
# needed for building with cachito
# https://github.com/release-engineering/cachito/blob/master/docs/pip.md
#
aiowsgi==0.7
alembic==1.3.3
aniso8601 @ git+https://github.com/DevTable/aniso8601-fake.git@bd7762c7dea0498706d3f57db60cd8a8af44ba90
APScheduler==3.6.3
attrs==19.3.0
aws-sam-translator==1.20.1
aws-xray-sdk==2.4.3
azure-core==1.8.0
azure-storage-blob==12.4.0
Authlib==1.0.0a1
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bencode.py==2.1.0
bintrees==2.1.0
bitmath==1.3.3.1
blinker==1.4
boto3==1.17.21
botocore==1.20.21
boto @ git+https://github.com/quay/boto.git@0041500f529547f592194c73c0a7d23275ed81c6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.3
cfn-lint==0.27.2
chardet==3.0.4
click==7.1.2
cnr-server @ git+https://github.com/quay/appr.git@58c88e4952e95935c0dd72d4a24b0c44f2249f5b
cookies==2.2.1
cryptography==3.4.6
DateTime==4.3
debtcollector==1.22.0
decorator==4.4.1
Deprecated==1.2.7
docker==4.1.0
docutils==0.15.2
dumb-init==1.2.2
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
fakeredis==1.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Testing @ git+https://github.com/jarus/flask-testing.git@17f19d7fee0e1e176703fc7cb04917a77913ba1a
funcparserlib==0.3.6
funcsigs==1.0.2
furl==2.1.0
future==0.18.2
futures==3.1.1
geoip2==3.0.0
gevent==1.4.0
gipc==1.0.1
greenlet==0.4.15
grpcio==1.30.0
gunicorn==20.0.4
hashids==1.2.0
hiredis==1.0.1
html5lib==1.0.1
httmock==1.3.0
httpretty==0.9.7
idna==2.8
importlib-metadata==1.4.0
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.0
jmespath==0.9.4
jsondiff==1.1.2
jsonpatch==1.24
jsonpath-rw==1.4.0
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.2.0
kafka-python==1.4.7
keystoneauth1==3.18.0
Mako==1.1.1
MarkupSafe==1.1.1
maxminddb==1.5.2
meld3==2.0.0
mixpanel==4.5.0
mock==3.0.5
mockldap @ git+https://github.com/quay/mockldap.git@4265554a3d89fe39bf05b18e91607bec3fcf215a
monotonic==1.5
msgpack==0.6.2
msrest==0.6.19
namedlist==1.7
ndg-httpsclient==0.5.1
netaddr==0.7.19
netifaces==0.10.9
oauthlib==3.1.0
orderedmultidict==1.0.1
oslo.config==7.0.0
oslo.i18n==3.25.1
oslo.serialization==2.29.2
oslo.utils==3.42.1
os-service-types==1.7.0
pathlib2==2.3.5
pathspec==0.7.0
pathvalidate==2.0.1
pbr==5.4.4
peewee==3.13.1
Pillow==7.2.0
pip==20.2.3
pipdeptree==1.0.0
ply==3.11
prometheus-client==0.7.1
protobuf==3.12.2
psutil==5.6.7
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
py-bitbucket @ git+https://github.com/quay/py-bitbucket.git@85301693ce3682f8e1244e90bd8a903181844bde
pycparser==2.20
pycryptodome==3.9.4
pycryptodomex==3.9.4
PyGithub==1.45
PyJWT==1.7.1
pymemcache==3.0.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
#PyPDF2==1.26.0
PyPDF2 @ https://files.pythonhosted.org/packages/b4/01/68fcc0d43daf4c6bdbc6b33cc3f77bda531c86b174cac56ef0ffdb96faab/PyPDF2-1.26.0.tar.gz#egg=PyPDF2&cachito_hash=sha256:e28f902f2f0a1603ea95ebe21dff311ef09be3d0f0ef29a3e44a932729564385
pyrsistent==0.15.7
pytest-runner==5.2
python-dateutil==2.8.1
python-editor==1.0.4
python-etcd @ git+https://github.com/DevTable/python-etcd.git@f1168cb02a2a8c83bec1108c6fcd8615ef463b14
python-gitlab==2.0.0
python-keystoneclient==3.22.0
python-ldap==3.2.0
python-magic==0.4.15
python-swiftclient==3.8.1
pytz==2019.3
PyYAML==5.3
raven==6.10.0
recaptcha2==0.1
redis==3.3.11
redlock==1.2.0
rehash==1.0.0
reportlab==3.5.34
requests==2.22.0
requests-aws4auth==0.9
requests-file==1.4.3
requests-oauthlib==1.3.0
responses==0.10.9
rfc3986==1.3.2
s3transfer==0.3.2
scandir==1.10.0
semantic-version==2.8.4
setuptools==50.3.0
setuptools-scm[toml]==4.1.2
six==1.14.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
stevedore==1.31.0
stringscore==0.1.0
stripe==2.42.0
supervisor==4.1.0
supervisor-logging==0.0.9
supervisor-stdout @ git+https://github.com/quay/supervisor-stdout.git@9bdf630d0d1b4a16cf8b60b96adafb1ed98e7380
tabulate==0.8.6
termcolor==1.1.0
text-unidecode==1.3
tldextract==2.2.2
toml==0.10.1
toposort==1.5
txaio==20.4.1
tzlocal==2.0.0
urllib3==1.25.8
waitress==1.4.2
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
Werkzeug==0.16.1
wheel==0.35.1
wrapt==1.11.2
xhtml2pdf==0.2.4
xmltodict==0.12.0
yapf==0.29.0
zipp==2.1.0
zope.interface==5.1.2
